Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/eval-scripts-setTimeout-blocked-expected.txt

CONSOLE MESSAGE: line 13: Refused to execute a script because 'unsafe-eval' does not appear in the script-src directive of the Content Security Policy.
ALERT: PASS
